egs/requirements.txt

anthropic
arxiv
mcp
pypdf2
python-dotenv
typing
uv

